network applet: Keep track of unmanaged networks (and listen to their (#7628)
state changes) but don't let them show up in the network lists, and
don't have them count when it comes to section titles.

This follows up on a an initial effort to hide unmanaged networks but
which didn't take into account the possibility of those networks later
becoming manageable.  The revert commit, link to the original, as well
as some explanation is here: a1281a5.