cs_themes.py: Defer loading the download page and chooser entries until after the page is loaded.
The spices/downloader stuff takes a long time to load, and ends up freezing
the ui a bit when the Themes page is selected.  Load it after any animations
have completed.

The chooser buttons are trying to all load at once.  Instead load them
sequentially and display a spinner until loading is complete.