windowManager: Abort the effect if we have no actor
We can also simply access window_type on meta_window instead of calling
get_window_type(), which saves us a trip to a C getter. This could
reduce a lot of function invocation in this file and in windowEffects.js
in a future PR. This was verified as working correctly by placing
`global.log(actor.meta_window.get_window_type() ===
actor.meta_window.window_type)`, which was always true.