Remove style-changed signal on iconBox
This calls get_length on every allocation. This style property will not
change until the theme changes, in which case we do a full refresh of
all actors on `theme-set` signal emission in applet.js. When the label
preferences are toggled between titleDisplay values 1 and 2-4, a full
refresh takes place as well.