panel-launchers: rework DND
Moves all applet DND code, except a backwards-compatibility stub,
onto a new class that handles DND for the actor in which we
actually keep the launcher actors. This prevents us having to
transform the coordinates from applet- to container-relative.

Uses separate behavior for "local" vs "remote" drags.
-If rearranging movement is done in place.
-If dragging in a new launcher a placeholder is used.
This simplifies drop index calculating quite a bit because
we can just use the index directly without having to
compensate because we're adding a fake index in the
local drag case by leaving the original actor and also
using a placeholder.

DND hover events are no longer processed fully while a
placeholder is being animated in or out. This can move
the pointer into a different index without much user
interaction and cause erratic results. Furthermore, in
certain panel layouts, the entire applet may shift slightly
under the cursor as an actor grows or shrinks, which adds
additional imprecision. Instead we now indicate no drop if
the placeholder has just been removed, or copy/move drop
when we are adding a placeholder.