gwl: Speed up initialization (#8092)
* gwl: Speed up initialization

Thumbnail menu and context menu construction now occur when they are
needed, and removes some left over timeout_add deferring from ITM.

Remove un-used properties

* Guard against more undefined hoverMenu cases

* Reduce thumbnail creation perf impact during window mapping

Thumbnail creation was only deferred for the case when the menu is
already open and a the middle mouse button is clicked on its
corresponding app button, for all other cases it can work synchronously.
This queues thumbnail creation on window added, which finishes on the
next menu open.

* Cache the theme node in style_length

We can reduce round-trips to C code this way.

* Simplify window check in AppList.windowAdded

* Don't re-create thumbnails when the orientation changes

* Remove style-changed signal on iconBox

This calls get_length on every allocation. This style property will not
change until the theme changes, in which case we do a full refresh of
all actors on `theme-set` signal emission in applet.js. When the label
preferences are toggled between titleDisplay values 1 and 2-4, a full
refresh takes place as well.

* Remove DND.LauncherDraggable _init call in AppGroup

This doesn't do anything.

* Fix prettier formatting