cinnamon-app: Get a fallback icon from Muffin for some window-backed apps
In muffin#399 icons are no longer created for every window because
Cinnamon does this most of the time. For window backed apps however,
Cinnamon needs to fallback to Muffin's icon. Instead of creating a
Muffin icon for every window, this will create them as needed, or in
this case, when we have a non-Xapp window-backed app.