menu/desktop editor: Run update-desktop-database on the user's applications folder when modifying a desktop entry.
If this does not run, the file is removed as a mime handler (unless it is
explicitly added again).  This command regenerates the mimeinfo.cache file
in the user's applications folder (typically ~/.local/share/applications).

This is being discussed upstream (the need to even do this, which I think
was not always the case): https://gitlab.gnome.org/GNOME/glib/issues/1610

Also, do a slight refactor which keeps a new desktop file being made until
the user actually wants to commit the change, not as soon as they click
'properties' in the menu editor.

Fixes #8149