gwl: Average the progress of all windows belonging to an app group button
The progress code was originally based on the Cinnamon window list's
code, but it didn't account for things GWL needs to account for, such as
multiple windows with progress sharing a single button.

This also splits the progress allocation into a stand-alone function so
we don't have to call `queue_relayout` when just the progress is
changing, and fixes the progress not showing after the applet or
Cinnamon is restarted.