Overview: don't close on key release just after opening from keyboard
The flag used to indicate if the key was already handled shouldn't be
initialized to false until a key is pressed, so that initial key release
events don't think they have not been handled in case the overview is
opened via a keyboard shortcut. This flag will be initialized to true/false
on key press anyways.

Fixes #8037