ShellApp: port to new GDBusActionGroup and GMenuProxy API
GDBusActionGroup and GMenuProxy are new objects in GIO 2.32 that
help with accessing menus and actions of remote applications.
This patch makes it possible for the shell to associate an
application with a dbus name and from that a GMenu, that will
be shown as the application menu.

https://bugzilla.gnome.org/show_bug.cgi?id=621203
https://github.com/GNOME/gnome-shell/commit/8764253861b2a065fa3e5b071a131831a91c3c5b

Application Menu: update for latest GMenu changes

GMenuProxy has been replaced by GDBusMenuModel, and the object path
has been moved (now needs to be retrieved from the AppMenu GApplication
property).
Update the test to prefix each action with "app." as documented,
and use a GtkApplicationWindow instead of a plain GtkWindow.

https://github.com/GNOME/gnome-shell/commit/5ad8080cb90dd3cfd247a85d5d68d29a1cf9c6ee

Another update for GLib API changes

GDBusActionGroup api has changed again, adapt to that.
Also, use a GActionMuxer to add the 'app.' prefix to actions,
instead of manually stripping it out of the action names.
In the future, the muxer will also contain per-window actions
with a 'win.' prefix.

https://github.com/GNOME/gnome-shell/commit/4aa1fe9ca2c133caa3e8e9bf6c0d9f14d30b3665

Add per-window actions

GTK+ also exports window-specific actions, by putting the object path
for the exported action group in the _DBUS_OBJECT_PATH X property.
We add this action group to the app's muxer with a 'win' prefix,
since that is what the exported menu expects. Whenever the focus
window changes, we update the window-specific actions of its
application, and emit notify::action-group to cause the app
menu to be updated.

https://github.com/GNOME/gnome-shell/commit/6c4e9d23f20585fe84c77d3ab6ce0637b380a0e9

shell-app: ensure there's always a muxer when setting actions on it

It's not guaranteed that the application DBus proxy appears before
we receive the first focus event from the toplevel window.
Ensure that the first method to access the action muxer creates it if
hasn't been created yet.

https://github.com/GNOME/gnome-shell/commit/7b9c9b2f7dabec457fb0e3f107dc80e24bf1885d

ShellApp: don't use the app proxy before it's created

The application proxy is created asynchrously after the dbus name
is registed. This means that when tracking the first window (and
therefore creating the first window GActionGroup) there is no
app proxy yet.

https://bugzilla.gnome.org/show_bug.cgi?id=633028
https://github.com/GNOME/gnome-shell/commit/f4d13b98012829f71bb50aa699ddac1671095e3e

ShellApp: Use unique name for menus

Otherwise since the auto-activate flag is set for dbus messages, we
may re-vivify applications that have DBus service files.

https://bugzilla.gnome.org/show_bug.cgi?id=667881
https://github.com/GNOME/gnome-shell/commit/b5be62cd1b0a050fdaff821fd9ef60e9c50dacd8

Adjust to gtk/mutter changes for Application API

https://bugzilla.gnome.org/show_bug.cgi?id=668118
https://github.com/GNOME/gnome-shell/commit/c5932c0f070a8969c7beafa420b2ebb58c3ee21