app-system: Lazily create ShellApps for apps we care about
Rather than create all ShellApps up-front, create them lazily. We really
had no reason to do this before as we were scanning GMenu to get all the
apps, but doing this can remove a need for get_all, which is slow and
memory-hungry.

https://github.com/GNOME/gnome-shell/commit/7890af16595dc61c9975f8f6261af099043c5199

app-system: Fix the keys in the id_to_app table

The ID that's passed to us isn't usable as a key to store, as it's
probably junk after the method returns. Use the app's storage for
the ID instead.

https://github.com/GNOME/gnome-shell/commit/dcb28aad2ab82566cd49b442766ed1aaf518ca84

app-system: Fix const warning

https://github.com/GNOME/gnome-shell/commit/a5619bc0a3f67fc52c98d4fbc8368193338881bd

app-system: Fix bad code

I had this locally, but forgot to amend before pushing apparently

https://github.com/GNOME/gnome-shell/commit/78343f483749c3637f873678bbe80ce04c39ed4f

app-system: Fix memory leak

The shell app takes a ref on the info.
https://github.com/GNOME/gnome-shell/commit/09fe31179aaf0ddf3e4786b61f8011431a5e72d5