menu applet: only unhighlight if highlighted
Previously, we would re-add the same app id to the _knownApps
array every time that button was activated. This isn't a big deal,
but could become a minor memory "leak" during longstanding sessions.

This makes sure we only unhighlight if we're highlighted, and
uses a set to simplify the _knownApp check. If we happen to get
highlighted more than once, the set prevents duplicates.