shell-app: Track the starting state correctly
Apps that are starting might have uninteresting windows like splash
screens pop up and then go away (like LibreOffice), even when
startup-notification hasn't completed yet. In those cases, we don't
want to transition the app back to stopped -- it should remain in
the running state.

https://github.com/GNOME/gnome-shell/commit/ff664fd1d89275a993e23ec7cf280fdf8c77001c

shell-app: Stay consistent about application states

When the last interesting window of an app-backed window is removed,
we'll transition it back to STOPPED, but we transition the state and
send out the signal before we clear the running state.

This means that any listeners to the state-changed signal might
encounter a window-backed app that has a running state, but no
windows. If they call, e.g. shell_app_get_name, while in this state,
they'll encounter an assertion fail.

https://github.com/GNOME/gnome-shell/commit/7424ee755a0cf5a21a0887bd72aa3e833b0892e1