xlet settings list widget: some fixes and improvements to the tree
This improves how data is displayed in the tree of an xlet settings
list widget in several ways:

1.) Boolean values are now displayed as a checkbox rather than the
values TRUE and FALSE.
2.) Icon type values are now displayed as an icon rather than text.
3.) Columns with a key-value options list now display the key
rather than the value. In most cases this makes more sense and it
brings it more in line with the combo box widget which is used in
the add/edit dialog.
4.) A new property 'align' was added to allow the alignment of the
content to be controlled. The settings example applet and the docs
were updated to reflect this change.