sound chooser widget: add all mimetypes option and improve default
This adds an option to the sound chooser widget to use any sound
file type - not just ogg/oga and wav (which are the ones supported
by the cinnamon sound system). The settings reference tutorial is
also updated to reflect the new option.

The default path for the file chooser dialog in this widget was also
set to /usr/share/sounds for cases where the file is not set or
doesn't exist as this is a much better default than ~/.