Various menu improvements (#8297)
* menu applet: simplify _select_category()

There's no need to call listApplications() since this is only
called with category names or null anyways.

* menu applet: only unhighlight if highlighted

Previously, we would re-add the same app id to the _knownApps
array every time that button was activated. This isn't a big deal,
but could become a minor memory "leak" during longstanding sessions.

This makes sure we only unhighlight if we're highlighted, and
uses a set to simplify the _knownApp check. If we happen to get
highlighted more than once, the set prevents duplicates.

* menu applet: simplify _listApplications()

Since we don't use this in _select_category() anymore we can remove
that logic, and in the one place it's used we can remove the empty
pattern check since we do that already inside the function and
return an empty array.

* menu applet: simplify _resizeApplicationsBox()

Runs 1-2ms faster generally, and seems to work just as well.

* menu applet: don't manually realize actors

from the docs:
"Actors are automatically realized, and nothing requires explicit
realization."

* menu applet: fix + simplify _displayButtons()

This makes sure all Transient and SearchProvider buttons are
cleaned up any time search re-runs or the category is changed.

The simplify part aggregates loops, and in the process switches
to Util.each() since it's a bit faster than Array.forEach().