TreeListWidgets.py: add ability to toggle booleans directly
Allows boolean items in the tree to be changed by checking or
unchecking the checkbox rather than having to edit the whole row.
The reason for this feature, and why the other settings are not
also editable is because a checkbox gives the impression that you
can click it, but the other settings types do not.