app-system: Reduce processing in cinnamon_app_system_lookup_app
This function gets called a lot, and seems to be continually failing to
find an app, and an app info object, causing it to always take the least
optimized path.

Just load all the apps on init and on install change like we've always
done, and get the heavy work out of the way first.