Revert "Backport appsys changes (#8295)"
This reverts commits f8fadd5be74196e3197323124b8eca657d9770c6
and 8717fa8c88461239d97584d0b0c0328c06036815.

We're not getting consensus on this within the team.

Jason is experiencing more input lag since the rebase of appsys.

The gains on loading aren't significant.

Stephen noted that we're now instantiating a new CMenu tree everytime
we need one, as opposed to a single one as it we did in the past.

Some of the commits also bring code which we don't need in Cinnamon.

We need to look at this again. This is getting reverted for now and
the PR will be reopened.