App: Don't listen to windows' user-time
This is overkill. The windows-changed signal isn't used anywhere,
it could be useful (in GWL for instance) but we don't want it
fired every single time a window is focused.

The only reason this was done was to sort the window list
in get_windows() and since nobody's listening to that signal,
it doesn't trigger any change to existing lists in any applets
out there.