Simplify Appsys (#8367)
* cinnamon-app: Cache a pointer to GDesktopAppInfo

* Don't crash on accesses to stale window-backed apps

The JS code could still be holding on to a reference to a window-backed app
after all windows have vanished. (For example, the dash queues an idle to
refetch apps and display them.) Avoid dying with an error message if we
attempt to activate or otherwise manipulate such a window.

https://bugzilla.gnome.org/show_bug.cgi?id=674799
https://github.com/GNOME/gnome-shell/commit/db297e7fdb08e80583c39bd6c59ddf05805f67ec

https://github.com/linuxmint/Cinnamon/pull/8295/commits/17c80919fc1769673f5ba28a5dcd0664a9cdcce3

* AppDisplay: don't show a "New Window" menu item if not possible

If the application reports itself as single window (through
an explicit indication in the desktop file or some heuristics),
not show a "New window" item that doesn't actually open a new window.

https://bugzilla.gnome.org/show_bug.cgi?id=722554

https://github.com/GNOME/gnome-shell/commit/3182aba744e7d64ecf71dfdbb5f5b1fa663325dc

Adapted
https://github.com/linuxmint/Cinnamon/pull/8295/commits/5adf8b34d6b70e8536709f244c435e3f3ad78461

* app-system: remove outdated comment

This parameter was removed before 3.2 but the comment stuck around.
https://github.com/GNOME/gnome-shell/commit/92915943304cb62b2b3a51e9141ec21a13053936

* Update (allow-none) annotations

The annotation has been deprecated in favor of (nullable) and/or
(optional).

* Doc fixes

https://bugzilla.gnome.org/show_bug.cgi?id=663277
https://github.com/GNOME/gnome-shell/commit/628e59894b2cf990dd2cb1e84be37ed09c6de486

* Remove obsolete app_setting tree and functions

These haven't been used for years and return NULL.

* Remove Faded icons

We don't use them.

* Appsys: Remove unused search functions

* Appsys: Remove lookup_app_for_path()

* Appsys: Remove lookup_app_by_tree_entry()

* Window tracker: Remove unused get_app_context()

* Appsys: Don't expose unused lookup_heuristic_basename

It's only used inside of appsys.

* App: remove compare_by_name()

unused.

* Fix warnings

* Appsys: Don't sort running apps

This is only used by the window tracker when looking for a match
on PID.. so we really don't need to sort lists all the time here.

* Appsys: Remove unused code

* App: Don't listen to windows' user-time

This is overkill. The windows-changed signal isn't used anywhere,
it could be useful (in GWL for instance) but we don't want it
fired every single time a window is focused.

The only reason this was done was to sort the window list
in get_windows() and since nobody's listening to that signal,
it doesn't trigger any change to existing lists in any applets
out there.