gwl: Check the cursor location after close click for good measure
When the menu size changes, it can leave the cursor outside the menu
bounds, and no leave event will be able to correct this situation where
the menu is dangling open and only closable upon hovering over it again.
This checks if the cursor is hovering over the menu 500ms after the last
close click, and closes it if not.