menu applet: only update label on orientation change
This doesn't actually work as is because it's just partially
recreating the menu, and leaves it in a broken state. As far
as I can tell we don't need to rebuild anything with it being
the same layout in every orientation.

This also removes the 'orientation' property because it already
exists on the prototype as '_orientation'. If we aren't modifying
it I don't see a reason to store it twice...