Remove maybeGet
This was mostly needed to shut up Spidermonkey warnings, but in most
cases Spidermonkey was being extra verbose because CJS/GJS configures
the environment that way. It is completely valid to just use a falsey
check before using an instanceof comparison, but even that is
unnecessary because <undefined value> instanceof {any object} will
always be false. In this case Spidermonkey's extra (forced) verbosity is
harmful and creating confusion.