Stop overriding Object.prototype.toString, move method to global scope
Native JS methods should not be overriden because it affects every part
of the application, and JS engines and libraries optimize under the
assumption that the base vanilla implementations work a certain way.

The try-catch will be slower and was removed, and if anything is
throwing here we should be handling errors in the paths least called,
not in core formatting methods.