keyboard: Silence a spidermonkey warning
Cjs-Message: 02:28:09.252: JS WARNING:
[/usr/share/cinnamon/js/ui/keyboard.js 509]: reference to undefined
property "extended_key"

This code assumes any given actor in a Clutter event function (which can
be propagated from practically any widget) has an extended_key property.
This is really silly that we have a nested event handler that doesn't
need anything exclusive to the keyboard class. Keep track of keyboard
visibility instead, and only use the contains method if visible. That is
essentially what was intended by this method because extended_key on an
actor would only be defined if the Key actor delegate class was
constructed (which only happens when the keyboard is enabled and
visible).