background effects: rework and turn on
This reworks, enables, and extends the previous background effect
facility that was turned off in the past as it was broken.

There are two possible background effects - bumpmap and blur. Both
are turned on by CSS, and can be layered on each other if desired.

example CSS:

background-bumpmap: url("/usr/share/cinnamon/bumpmaps/hex.png");
background-blur: 1px;

Note that the size of the blur is treated as an iteration count. So
background-blur: 1px is the basic effect, while background-blur: 2px
applies the effect again on top of itself, and so on. The basic blur
is relatively mild, twice is stronger, anything above 3 is overkill,
IMO.

Speeds seem fine with no noticeable speed penalty unless a large blur
count is put in for test purposes.

The blur effect is probably the more useful. The bumpmap effect looks
dramatic on a background, but rubbish over something like a text file
being edited.

The effect works by snapshotting the screen underneath an actor,
applying the effect to it, and then drawing it out as the lowest
layer in the process of drawing an actor. There are no restrictions
as to what can be layered on it, though clearly you won't want to put
something without transparency over it or the effect would be hidden.