layout: Cache the rects and struts as JS objects, only update on change
If a function is called enough, it can be faster to stay in JS and not
call back C bindings to avoid context switching.

- Converted meta_rectangle_intersect to JS
- Cache the rects and struts as JS objects, so we don't hold a bunch of
boxed refs implicitly. The caching lets us detect strut/rect change, in
which case the C functions are called. The user will be interacting with
applications, not the Cinnamon UI most of the time, so this is an
important improvement.
- We don't have per-workspace panels, so there is no need to
re-calculate everything on workspace switch, number change.

Seeing lower CPU and memory usage from this commit.