PIL: remove unnnecessary version check and code fork
The Image.VERSION variable was deprecated some time ago and is now gone
entirely. But its only use was to determine whether we were using at
least the latest version of PIL, or any version of its fork, Pillow. And
PIL was last developed in 2009 and does not support python3 at all, so
we are guaranteed to be using Pillow.

Moreover, the check does not really matter, as Pillow is guaranteed to
load any image it opens, without the user doing so manually.

Fixes #8495