Some adjustments for the new version of cmenu
The majority of these changes are due to the change in prefix from gmenu_* to cmenu_*, though support for inlining and separators were also removed since we don't use those in Cinnamon anyway. CMenu now also flattens the menu structure and performs a sort, so we don't need to do that in Cinnamon any more.