layout: Misc. clean up (#8465)
* layout: Don't block _queueUpdateRegions

Signal emission order is not guarunteed, but apparently is important -
we need the function that calls back into muffin to have higher
priority, otherwise bottlenecks compositor perf.

* Fix up _windowsRestacked

- isPopupMetaWindow is only used in one place, doesn't need to be a
function call

- this._isPopupMenuVisible was undefined, and setup a condition in
_queueUpdateRegions that could never be false.

Remove unused variable

* Remove Lang.bind usage, lower priority of _windowsRestacked

* layout: Throttle global.sync_pointer to one call per second

Restacks can propagate for any number of reasons, even from windows that
muffin does not composite. On my machine this restacked signal is
emitted a few times per second, even on a fresh session with no windows
open. This makes sure this heavy X round trip doesn't block the
compositor.

* Use global.display.popup_window_visible

* cinnamon-global: Use meta_display_get_pointer

This lets us avoid going through a lot of GDK class abstraction to get
values from XQueryPointer.

* Use length caching for loops

* Pass monitor object directly to meta_rectangle_new

CJS will ignore the getter.

* Don't throttle global.sync_pointer

Noticed some tray app icons not disappearing after closing. Since we're
limiting this to a changed state we don't need the throttling now.

* Increase updateRegions priority

* Lower restacked priority to 1000

* layout: Use a more appropriate signal for stack detection

The restacked signal is getting emitted too frequently. We can preserve
the intent behind `_windowRestacked` by just listening to
notify::focus-window. Even if a window is minimized, the focus window
changes. The topmost window should always have focus, and any window
operation affecting it will propagate to this signal's emission.

* layout: Cache the rects and struts as JS objects, only update on change

If a function is called enough, it can be faster to stay in JS and not
call back C bindings to avoid context switching.

- Converted meta_rectangle_intersect to JS
- Cache the rects and struts as JS objects, so we don't hold a bunch of
boxed refs implicitly. The caching lets us detect strut/rect change, in
which case the C functions are called. The user will be interacting with
applications, not the Cinnamon UI most of the time, so this is an
important improvement.
- We don't have per-workspace panels, so there is no need to
re-calculate everything on workspace switch, number change.

Seeing lower CPU and memory usage from this commit.

* Revert accidental deletion of comment

* Update struts after panel hide, remove

This happens because the trackedActors array length is the only thing
that is changing (the panel being removed), and that is not being
checked in updateRegions. Since that method is called constantly, this
just overrides the changed state instead of adding more logic to it.

* Update regions on PopupMenu close

Fixes unclickable areas where a GWL thumbnail used to be.