workspace-switcher applet: Fix updating the focus window position and size at startup, and ensure the focus window starts out on top of the graph button.
The repaint function for workspace graph buttons was relying on the
last user time for each window to sort them, and this seems to be
not entirely reliable.  Just make sure to always paint the focused
window's rectangle last.

The _createButtons methods was disconnecting all signals before re-
connecting the focus-changed handler.  If _onFocusChanged returned
early (because the focus window stayed the same) the move and resize
handlers wouldn't be reconnected.  Since these three handlers are the
only ones the signal manager is keeping track of, just selectively
disconnect instead of disconnecting all.