menu-applet: use displayButtons for initial item visibility
Otherwise, we may have places or recents mixed in if they have
changed since the menu was last opened, and since we are iterating
all menu items in displayButtons and not resuming where we left off
with INITIAL_BUTTION_LOAD, we only initially show 15 items now.

This also gets rid of hiding all apps on menu close. We always open to
all apps, so theres no reason to hide them all on close.