windowEffects: Don't use clones for the close effect
It appears MetaWindowActor is being disposed during the effect
because once switched over to applying the effect directly, we need a
finalized checked in Effect.prototype._end. This is breaking Clutter's
ability to opacify the clone for some reason. We can apply the effect
directly as a fix since there weren't any issues observed with the close
effect in particular.