windowEffects: Fix opacity during the close effect, stacking during the minimize effect (#8596)
* windowEffects: Don't use clones for the close effect

It appears MetaWindowActor is being disposed during the effect
because once switched over to applying the effect directly, we need a
finalized checked in Effect.prototype._end. This is breaking Clutter's
ability to opacify the clone for some reason. We can apply the effect
directly as a fix since there weren't any issues observed with the close
effect in particular.

* Remove else operator

* windowEffects: Apply the minimize effect on MetaWindowActor directly

This is a solution for the discussed issue in
https://github.com/linuxmint/cinnamon/pull/8556. It is tricky getting
the clones to be layered correctly if the animation does not start with
the focused window. For effects where the focus window is implied, e.g.
map, unminimize, we can continue using clones.