st-label: Move St-specific shadow processing out of the paint vfunc (#8588)
* st-label: Revert changes from 37d75d48e4f0214e498dd559ee257946ad3fed6f

* Avoid a function call during paint

Also not NULL'ing it because it gets re-assigned.

* Simplify expressions

This would need strict comparison if we were adding an incompatible type
to the expression.

* Remove another function call from StLabel's paint vfunc

* Move StWidgetPrivate to st-private.h so it can be used in st-label.c

* st-label: Move St-specific shadow processing out of the paint vfunc

We are fetching the theme node and shadow object on paint - both heavy
calls. That is our only way of knowing if we should process text
shadows. Very few labels actually make use of this feature in Cinnamon,
so move it to the allocate vfunc, which will get called less but after
any style change. We can only know from both the style change and
allocate vfuncs if the style has actually changed by checking
`is_style_dirty` on StWidgetPrivate. The alternative is to change the
public API which will break things. This is better long-term because
many internal St methods are reusing StWidget's public API when they
don't need to.

This also removes the type cast instance checks in touched methods
because buried in the layers of macros is a call to
`g_type_check_instance_cast` in GLib, which will incur a performance
penalty. We don't need to worry about type checking at runtime in
internal vfuncs.