Xlets removal confirmation dialogs (#8636)
* [applet.js] Add confirmation dialog when removing an applet from its context menu

This commit adds a confirmation dialog when trying to remove an applet from a panel when its "Remove" context menu item is activated. If the Control key is pressed when pressing "Remove", the applet is removed without confirmation.

Closes #8634

* [desklet.js] Add confirmation dialog when removing a desklet from its context menu

This commit adds a confirmation dialog when trying to remove a desklet from the desktop when its "Remove" context menu item is activated. If the Control key is pressed when pressing "Remove", the desklet is removed without confirmation.