remove js-based automated perf testing code
`system.gc()` hasn't worked for a while, and newer versions of
mozjs remove support for `StopIterator`. As is, this code errors
on the past few versions, and will break more in the future.
Rather than figure out how to rework the code to work as intended,
it's being removed as it seems it isn't used enough for anyone to
maintain.