st-theme-node.c: Allow parsing of 'symbolic' (and -warning, -error, and -success) as a color.
This works its way up thru the css hierarchy until it finds any
corresponding color, warning-color, error-color, and success-color,
respectively, and uses that.

It's similar to how the 'inherit' keyword works, but it allows you
to be explicit about what color you want applied to what themed element.

The means of establishing the symbolic colors is pre-existing (for
StIcons).  This only adds the capability of using these colors within
the theme.