runDialog.js: Go back to using the original spawnCommandLineAsync, as well as the original code using it in runDialog.
This was rendering _showError() useless, as the new piped spawn
function forced you to wait until the async run was complete before
getting an error back.  By this time the dialog was closed, so you
never knew if something failed to execute or not.

The focus handling bits were also breaking the right-click context
menu (cinnamonEntry) as well as breaking alt-f2 if the mouse pointer
happened to be over a panel.

The piped async function is kept, but renamed.

Reverts some of fa9a3c63d4da8